refactor(admin): Q9.e — extract admin/observability.rs (GAR-473)
Extracts Phase 6 Observability/UI handlers from admin/handlers.rs into new
admin/observability.rs (~230 LOC): admin_logs, admin_metrics, admin_prometheus,
admin_alerts, list_themes, get_layout_preferences, list_templates, about.

handlers.rs: 2326 → 2103 LOC (−223 LOC). Zero behavior change.
Re-exported via `pub use super::observability::*` in handlers.rs so
routes.rs paths (`handlers::admin_logs`, etc.) keep resolving unchanged.

Follows exact pattern of Q9.a/b/c/d (GAR-439/470/471/472).
